+++ app/controllers/custom_products_controller.rb
@@ -3,23 +3,27 @@ class CustomProductsController < ApplicationController
   before_action :set_breadcrumb, except: [:show]
 
   def index
-    @page_title = I18n.t('headings.setups')
-    @custom_products = current_user.possessions
-                                   .where.not(custom_product_id: nil)
-                                   .map { |possession| CustomProductPresenter.new(possession) }
-                                   .sort_by(&:name)
+    @page_title = I18n.t('headings.custom_products')
+
+    @custom_products = current_user.custom_products
+                                   .all
+                                   .includes([image_attachment: [:blob]])
+                                   .includes([:sub_categories])
+                                   .order(:name).map do |custom_product|
+      CustomProductPresenter.new(custom_product)
+    end
   end
 
   def show
-    @user = User.find_by('lower(user_name) = ?', (params[:user_id].presence || params[:id]).downcase)
+    @user = User.find_by('lower(user_name) = ?', params[:user_id].downcase)
+
     possession = @user.possessions.find_by(custom_product_id: params[:id])
+    @possession = CustomProductPossessionPresenter.new(possession) if possession
 
-    return render 'not_found', status: :not_found if possession.nil?
-
-    @possession = CustomProductPresenter.new(possession)
+    @custom_product = CustomProductPresenter.new(@user.custom_products.find(params[:id]))
 
     unless current_user == @user
-      redirect_path = get_redirect_if_unauthorized(@user, @possession)
+      redirect_path = get_redirect_if_unauthorized(@user, @custom_product)
       return redirect_to redirect_path if redirect_path
     end
 
@@ -27,8 +31,8 @@ class CustomProductsController < ApplicationController
 
     add_breadcrumb I18n.t('users'), users_path
     add_breadcrumb @user.user_name, user_path(id: @user.user_name.downcase)
-    add_breadcrumb @possession.name
-    @page_title = @possession.name
+    add_breadcrumb @custom_product.display_name
+    @page_title = @custom_product.display_name
   end
 
   def new
@@ -39,10 +43,15 @@ class CustomProductsController < ApplicationController
   end
 
   def create
-    @custom_product = CustomProduct.new(custom_product_params)
-    possession = Possession.new(user: current_user, custom_product: @custom_product)
+    @custom_product = current_user.custom_products.new custom_product_params
 
-    if @custom_product.save && possession.save
+    if @custom_product.save
+      possession = Possession.new(
+        user: current_user,
+        custom_product: @custom_product,
+      )
+      possession.save
+
       flash[:notice] = I18n.t(
         'custom_products.created',
         name: @custom_product.name
@@ -69,6 +78,8 @@ class CustomProductsController < ApplicationController
 
   def update
     @custom_product = current_user.custom_products.find(params[:id])
+
+    @custom_product.image.purge if params[:custom_product][:delete_image]
 
     if @custom_product.update(custom_product_params)
       flash[:notice] = I18n.t(
@@ -101,7 +112,7 @@ class CustomProductsController < ApplicationController
   end
 
   def custom_product_params
-    params.require(:custom_product).permit(:name, :description, sub_category_ids: [])
+    params.require(:custom_product).permit(:name, :description, :image, sub_category_ids: [])
   end
 
   def get_redirect_if_unauthorized(user, possession)
@@ -113,7 +124,7 @@ class CustomProductsController < ApplicationController
     # if the visited profile is not visible to anyone and the visiting user is a different user
     return root_url if user.hidden? && current_user != user
 
-    redir = user_custom_product_path(user_id: user.user_name.downcase, id: possession.id)
+    redir = user_custom_product_path(id: possession.id, user_id: user.user_name.downcase)
     new_user_session_url(redirect: URI.parse(redir).path)
   end
 end