+++ app/presenters/current_possession_presenter.rb
@@ -1,0 +1,9 @@
+class CurrentPossessionPresenter < PossessionPresenter
+  def delete_button_label
+    I18n.t('remove_product.label')
+  end
+
+  def delete_confirm_msg
+    I18n.t('remove_product.confirm', name: @object.display_name)
+  end
+end