+++ app/presenters/custom_product_setup_possession_presenter.rb
@@ -1,4 +1,4 @@
-class SetupItemPresenter < ItemPresenter
+class CustomProductSetupPossessionPresenter < CustomProductPossessionPresenter
   def initialize(object, setup)
     super(object)