+++ db/migrate/20240812070235_add_false_as_default_for_prev_owned_on_possessions.rb
@@ -1,0 +1,5 @@
+class AddFalseAsDefaultForPrevOwnedOnPossessions < ActiveRecord::Migration[7.1]
+  def change
+    change_column :possessions, :prev_owned, :boolean, default: false
+  end
+end